Add ability to set the background color.
The pixmap in TurtleGraphicsScene is now transparent, and the background
color is set in the background color of the QGraphicsView. This allows
the background color to be changed on-the-fly without needing to clear
the current drawing.

The new command exposed by the C++ application to the Lua scripts is
set_background_color(r,g,b) which takes three numbers in the range [0,255].

The bgcolor(r,g,b) function is implemented in the Lua script as a user-friendly
wrapper around set_background_color. It can be called in different ways, similar
to the color() function:
   bgcolor(255,0,0)
   bgcolor(red)