Replace QGraphicsView with a custom widget.
The previous code using QGraphicsView had terrible performance with large pixmaps
(e.g. 8192x8192 and above) due to the pixmap being copied on each UI update.

This new implementation replaces the use of QGraphicsView for displaying the
drawing with a custom widget. The widget still renders the turtle drawings to
a pixmap like before, but now the widget's paint events are intercepted to draw
to the UI directly from a portion of the pixmap, which is very efficient.

The custom widget is contained in a QScrollArea to allow viewing different parts
of the canvas.

The disadvantage to this approach is that QGraphicsView supported dragging,
which is no longer working in this commmit.