Snap turtle position to nearest integer coordinates.
QPainter appears to discard the fractional part of a float position when drawing
some lines. For example, the coordinate 4.999999 appears to be snapped to 4.
This causes unexpected artifacts in some scripts, for example in the following
script most lines will have a 1px gap, however some lines will be drawn without
the expected 1px gap:
   for n=1,1000,1 do fd(n) rt(90) end

This change ensures that the turtle's position is always snapped to the nearest
integer coordinate. For example, the coordinate 4.999999 is rounded to 5, and
4.24 is rounded to 4. Likewise, -4.99999 is rounded to -5 and -4.24 is rounded
to -4.

This snapping logic is configurable in the Turtle class in turtle.lua, but
currently there is no user command to control this.