Refactor C++/Lua API
The new API puts the C++ functions that are available to lua in a
global table to control the UI from within the lua script (e.g.
drawing to the canvas).

The new API is:
   _ui.print()
   _ui.canvas.drawline()
   _ui.canvas.drawarc()
   _ui.canvas.clear()
   _ui.canvas.setbackgroundcolor()
   _ui.canvas.backgroundcolor()

The CommandRunner class has been refactored into the ScriptRunner class,
and the C++ functions registered with Lua are merged into this class
for better integration.