Fix rendering artifacts without snapping turtle position.
Rendering artifacts with lines was previously fixed by snapping the
turtle's position to the nearest integer coordinate. However, this
does not work well with scripts dealing with precision turtle positions,
as the position is modified.

This fix allows the turtle to move completely freely without snapping
its position to the nearest integer coordinates, but still solves
the rendering artifact problem.

The fix for the artifact rendering is moved from turtle.lua to
TurtleCanvasGraphicsItem::drawLine() in C++, which performs rounding
of the line's coordinates prior to drawing. This rounding only occurs
when AA is disabled, since rendering works correctly with AA enabled.