Add options for saving canvas images.
This change adds a new dialog which appears when a canvas is saved.
The dialog contains two options:
   1. Fit to drawing
   2. Transparent backround

Option #1 fits the size of the output image to the canvas rect that
has actually been drawn to.

Option #2 selects whether or not the background is transparent in
the output image. If the image is not transparent then the current
screen color is used.