Fix unresponsive UI when the script prints too many messages.
This change adjusts the model for passing messages to the UI from a "push"
model to a "pull" model. Previously, ScriptRunner emitted a (queued) signal
containing the string printed by the script calling _ui.print(), which
MainWindow would receive and print to the UI in a text edit. However, if
_ui.print() is called too frequently then the UI became unresponsive due to
processing many queued signals.

The new model for transferring messages to the UI is a "pull" model. The
ScriptRunner's thread emits a (still queued) signal (scriptMessageReceived()),
which the UI must then call ScriptRunner::pendingMessage() to get the message.

If the Lua script tries to print another message before the UI has been able
to call ScriptRunner::pendingMessage() then the ScriptRunner's thread waits
for the pendingMessage() to be called by the UI (to "pull" the previous message)
before sending the next one.

This ensures that there is no more than 1 scriptMessageReceived() signal pending
in the MainWindow's event queue, and the UI is fully responsive in all cases.

The Lua script used to produce the unresponsive UI initially (and used to test
it is fixed with this change) is:
    while true do
        print("hello")
    end